# PERF: Implement RangeIndex min/max using RangeIndex properties