--- a/doc/source/api.rst
+++ b/doc/source/api.rst
@@ -1416,49 +1416,19 @@ Selecting
    Index.slice_indexer
    Index.slice_locs
 
-.. _api.int64index:
+.. _api.numericindex:
 
-Int64Index
-----------
-
-.. autosummary::
-   :toctree: generated/
-   :template: autosummary/class_without_autosummary.rst
-
-   Int64Index
-
-.. _api.uint64index:
-
-UInt64Index
------------
-
-.. autosummary::
-   :toctree: generated/
-   :template: autosummary/class_without_autosummary.rst
-
-   UInt64Index
-
-.. _api.float64index:
-
-Float64Index
-------------
-
-.. autosummary::
-   :toctree: generated/
-   :template: autosummary/class_without_autosummary.rst
-
-   Float64Index
-
-.. _api.rangeindex:
-
-RangeIndex
-----------
+Numeric Index
+-------------
 
 .. autosummary::
    :toctree: generated/
    :template: autosummary/class_without_autosummary.rst
 
    RangeIndex
+   Int64Index
+   UInt64Index
+   Float64Index
 
 .. _api.categoricalindex: