--- a/pandas/tests/indexes/test_range.py
+++ b/pandas/tests/indexes/test_range.py
@@ -10,7 +10,7 @@ from pandas.compat import range, u, PY3
 
 import numpy as np
 
-from pandas import (notna, Series, Index, Float64Index,
+from pandas import (isna, notna, Series, Index, Float64Index,
                     Int64Index, RangeIndex)
 
 import pandas.util.testing as tm
@@ -994,3 +994,22 @@ class TestRangeIndex(Numeric):
                 # Append single item rather than list
                 result2 = indices[0].append(indices[1])
                 tm.assert_index_equal(result2, expected, exact=True)
+
+    def test_max_min(self):
+        params = [(0, 400, 3), (500, 0, -6), (-10**6, 10**6, 4),
+                  (10**6, -10**6, -4), (0, 10, 20)]
+        for start, stop, step in params:
+            idx = RangeIndex(start, stop, step)
+
+            expected = idx._int64index.max()
+            result = idx.max()
+            assert result == expected
+
+            expected = idx._int64index.min()
+            result = idx.min()
+            assert result == expected
+
+        # empty
+        idx = RangeIndex(0, 1, -1)
+        assert isna(idx.max())
+        assert isna(idx.min())