- [X] closes #17607
- [X] tests added / passed
- [X] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [X] whatsnew entry

Benchmarks are essentially the same as what's referenced in the issue.  Maybe a few ns slower since I combined logic into a helper function vs having identical code (other than `elif`) for both min/max, which was the case I did the benchmarks posted in the issue.  Updated benchmarks:

```
      before           after         ratio
     [fedf9228]       [4cb0de8d]
-        25.4±0ms      1.77±0.06μs     0.00  index_object.Range.time_min
-      26.0±0.9ms      1.71±0.06μs     0.00  index_object.Range.time_max
-        25.4±0ms         1.51±0μs     0.00  index_object.Range.time_max_trivial
-        25.4±0ms         1.30±0μs     0.00  index_object.Range.time_min_trivial

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
```