@topper-123 2017-09-21
I'm wondering it something similar, but more general, couldn't be done that works with both indices and series using ``is_monotonic_*``. 

E.g.:

```python
def max(self):
    if self.is_monotonic_increasing:
        return self[-1]
    elif self.is_monotonic_decreasing:
        return self[0]
    else:
        return nanops.nanmax(self.values)
```

Seems to me this should work. What do you think, have I missed something nonobvious?