@jschendel 2017-09-21
@topper-123 : I think that will work in general, but I wonder if there are cases where the additional overhead of computing `is_monotonic_*` could outweigh performance gains.  Would be interesting to look into.